restructure mean module; add harmonic/geometric
The functionality of the ci is now captured in a trait `MeanCI` which must
be imported together with the concrete calculation type. Each kind of
c.i. is represented by its own structure: `Arithmetic`, `Geometric`,
`Harmonic`. Each computes the confidence intervals around the
corresponding mean.

The restructuring replaces the single function ci with a generic
(private) function `ci_with_transforms` which takes validity and
transformation closures as parameters. The private function is
used to produce the c.i. for each of the three kinds of means.