fix!: use wilson score for quantile::ci
BREAKING CHANGE:
the intervals computed by the Wilson score are somewhat different from
those computed with the z-value
In addition, the Wilson score method tests for statistical significance,
which was not the case with the computations based on the z-value.
Consequently, some quantiles that could be computed before are no longer
accepted due to the lack of statistical significance.