fix: switch to z-value for large populations
When the number of degrees of freedom reaches values around 15 millions,
the inverse_cdf of the t-distribution (in statrs) appears to hang.
The computation of interval bounds now switches to the z-value of the
normal distribution when the population is over a constant.
The threshold is currently set to 100_000 which is well above the common
folklore value of 30 found in most textbooks.
A higher value is not recommended because the inverse_cdf of the student
distribution seem to yield inconsistent results well before it hangs.