refactor: move methods from trait to impl
move the main methods to the impl of each struct and overload in the
implementation of the traits `StatisticsOps` and `MeanCI`.
This allows the methods to be accessible without importing the trait
and the trait is now only useful for rudimentary polymorphism.
This also brings the documented methods to the forefront.