docs: combine README and root doc
This combines the README with the doc for the root of the crate.
This is an entire rewrite of the documentation, focussing on the simpler
interface rather than the flexible interface using statistics objects.
While the latter is more flexible, most scenarios are expected to only
need the simpler one.
The documentation is more consistent with recent changes.